move channels and messages to channels domain
