+++ server/src/controllers/index.ts
@@ -1,9 +1,3 @@
-import ChannelControllerHTTP from "./channelController.js";
-import MessageControllerHTTP from "./messageController.js";
 import RelationshipControllerHTTP from "./relationshipController.js";
 
-export {
-  ChannelControllerHTTP,
-  MessageControllerHTTP,
-  RelationshipControllerHTTP,
-};
+export { RelationshipControllerHTTP };
