+++ server/src/domains/auth/index.ts
@@ -1,2 +1,3 @@
 export { AuthControllerHTTP } from "./controllers/http/auth.js";
 export { AuthService } from "./services/auth.js";
+
