+++ server/src/domains/channels/index.ts
@@ -1,0 +1,4 @@
+export { ChannelControllerHTTP } from "./controllers/channel.js";
+export { MessageControllerHTTP } from "./controllers/message.js";
+export { ChannelModel } from "./models/channel.js";
+export { MessageService } from "./services/message.js";
