+++ server/src/domains/channels/interfaces/channel.ts
@@ -1,0 +1,7 @@
+import type { Message } from "../interfaces/message.js";
+
+export type Channel = {
+  id: string;
+  messages: Message[];
+  participants: string[];
+};
