+++ server/src/domains/channels/interfaces/message.ts
@@ -1,0 +1,6 @@
+export type Message = {
+  _id: string;
+  senderId: string;
+  content: string;
+  timestamp: number;
+};
