+++ server/src/domains/users/index.ts
@@ -1,7 +1,7 @@
-export { UserControllerHTTP } from "./controllers/http/user.js";
 export { StatusControllerHTTP } from "./controllers/http/status.js";
-export { UserModel } from "./models/user.js";
-export { UserService } from "./services/user.js";
-export { StatusService } from "./services/status.js";
+export { UserControllerHTTP } from "./controllers/http/user.js";
 export * from "./entities/user.js";
 export type { User, UserProfile, UserStatus } from "./interfaces/user.js";
+export { UserModel } from "./models/user.js";
+export { StatusService } from "./services/status.js";
+export { UserService } from "./services/user.js";
