+++ server/src/models/index.ts
@@ -1,5 +1,4 @@
-import ChannelModel from "./channelModel.js";
 import FriendInviteModel from "./friendInviteModel.js";
 import RelationModel from "./relationModel.js";
 
-export { ChannelModel, FriendInviteModel, RelationModel };
+export { FriendInviteModel, RelationModel };
