+++ server/src/services/index.ts
@@ -1,5 +1,4 @@
-import MessageService from "./messageService.js";
 import RelationshipService from "./relationshipService.js";
 import NotificationService from "./notificationService.js";
 
-export { MessageService, RelationshipService, NotificationService };
+export { RelationshipService, NotificationService };
