+++ server/src/tests/mocks/models/mockChannelModel.ts
@@ -1,7 +1,7 @@
-import { Ok } from "resultat";
 import { ObjectId, type Db } from "mongodb";
-import type { IChannelModel } from "../../../interfaces/channelModelInterface.js";
-import { type Message } from "../../../entities/message.js";
+import { Ok } from "resultat";
+import type { Message } from "../../../domains/channels/interfaces/message.js";
+import type { IChannelModel } from "../../../domains/channels/models/channel.interface.js";
 
 class MockChannelModel implements IChannelModel {
   constructor(db: Db) {}
